Refactor PEC2Point into own module
