+++ grumplestiltskin.cabal
@@ -66,6 +66,7 @@ library
   exposed-modules:
     Grumplestiltskin.Degree1.EllipticCurve
     Grumplestiltskin.Degree1.Galois
+    Grumplestiltskin.Degree2.AffinePoint
     Grumplestiltskin.Degree2.Element
     Grumplestiltskin.Degree2.EllipticCurve
     Grumplestiltskin.Degree2.Galois
